Fix image build when no packages specified
If building an image with no packages specified, the first element in
the list is interpreted as a package. Often this element is the OS
element, in which case the build will fail early. In other cases it may
fail when a package with the name of the first element is not found.

This change fixes the issue by only specifying the -p argument when the
list of packages is defined and non-empty.

Fixes: #26