Set loop label when iterating over os_images_list
This can potentially leak sensitive information e.g a devuser password.
Using the whole dictionary as the loop label is also very verbose.
You can easily map back to the definition using the name. It should be
noted that the `default` stdout callback plugin will still print the
image definition on failure[1]. You can get round this by using an
alternative callback plugin. You can't really set the loop label
in the callback plugin as you do not know which field to use for the label.

[1] https://serverfault.com/questions/1059530/how-to-not-print-items-in-an-ansible-loop-error-without-no-log